+++ activesupport/CHANGELOG.md
@@ -1,3 +1,13 @@
+*   Fix `ActiveSupport::TimeZone#strptime`. Now raises `ArgumentError` when the
+    given time doesn't match the format. The error is the same as the one given
+    by Ruby's `Date.strptime`. Previously it raised
+    `NoMethodError: undefined method empty? for nil:NilClass.` due to a bug.
+
+    Fixes #25701.
+
+    *John Gesimondo*
+
+
 ## Rails 5.0.0 (June 30, 2016) ##
 
 *   Support parsing JSON time in ISO8601 local time strings in