+++ activesupport/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Rails 4.2.7.rc1 (June 30, 2016) ##
+## Rails 4.2.7 (July 12, 2016) ##
 
 *   Fixed `ActiveSupport::Logger.broadcast` so that calls to `#silence` now
     properly delegate to all loggers. Silencing now properly suppresses logging